chore: moves files in tests/unit to tests (#417)
