+++ .github/workflows/build.yml
@@ -68,7 +68,7 @@ jobs:
         run: poetry run pyright
       - name: Run Unit Tests
     #   Would use make, but want cov report in xml format
-        run: poetry run pytest -v --cov=sae_lens/ --cov-report=term-missing --cov-branch tests/unit --cov-report=xml
+        run: poetry run pytest -v --cov=sae_lens/ --cov-report=term-missing --cov-branch tests --cov-report=xml
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v4.0.1
         with:
