+++ makefile
@@ -10,19 +10,12 @@ check-type:
 	poetry run pyright .
 
 test:
-	make unit-test
-	make acceptance-test
-
-unit-test:
-	poetry run pytest -v --cov=sae_lens/ --cov-report=term-missing --cov-branch tests/unit
-
-acceptance-test:
-	poetry run pytest -v --cov=sae_lens/ --cov-report=term-missing --cov-branch tests/acceptance
+	poetry run pytest -v --cov=sae_lens/ --cov-report=term-missing --cov-branch tests
 
 check-ci:
 	make check-format
 	make check-type
-	make unit-test
+	make test
 
 docstring-coverage:
 	poetry run docstr-coverage sae_lens --skip-file-doc
