+++ pyproject.toml
@@ -68,7 +68,7 @@ ignore = ["E203", "E501", "E731", "F722", "E741", "F821", "F403", "ARG002"]
 select = ["UP", "TID", "I", "F", "E", "ARG", "SIM", "RET", "LOG", "T20"]
 
 [tool.ruff.lint.per-file-ignores]
-"tests/benchmark/*" = ["T20"]
+"benchmark/*" = ["T20"]
 "scripts/*" = ["T20"]
 
 [tool.pyright]
