+++ tests/training/test_config.py
