+++ tests/training/test_upload_saes_to_huggingface.py
@@ -12,7 +12,7 @@ from sae_lens.training.upload_saes_to_huggingface import (
     _repo_file_exists,
     _validate_sae_path,
 )
-from tests.unit.helpers import build_sae_cfg
+from tests.helpers import build_sae_cfg
 
 
 def test_create_default_readme():
